Move function to helpers.py
