+++ tests/helpers.py
@@ -2205,3 +2205,31 @@ def validate_pod_oomkilled(
         )
 
     return rc
+
+
+def validate_pods_are_running_and_not_restarted(
+    pod_name, pod_restart_count, namespace
+):
+    """
+    Validate given pod is in running state and not restarted or re-spinned
+
+    Args:
+        pod_name (str): Name of the pod
+        pod_restart_count (int): Restart count of pod
+        namespace (str): Namespace of the pod
+
+    Returns:
+        bool : True if pod is in running state and restart
+               count matches the previous one
+
+    """
+    ocp_obj = ocp.OCP(kind=constants.POD, namespace=namespace)
+    pod_obj = ocp_obj.get(resource_name=pod_name)
+    restart_count = pod_obj.get('status').get('containerStatuses')[0].get('restartCount')
+    pod_state = pod_obj.get('status').get('phase')
+    if pod_state == 'Running' and restart_count == pod_restart_count:
+        logger.info("Pod is running state and restart count matches with previous one")
+        return True
+    logger.error(f"Pod is in {pod_state} state and restart count of pod {restart_count}")
+    logger.info(f"{pod_obj}")
+    return False
